feat(renderer): expand distant-world capacities
Relocate GTA's visible entity and visible LOD pointer arrays from 1000 to 8192 entries, and raise the streaming RwObject instance list from 2500 to 30000 using full-width allocation immediates.

Add renderer high-water telemetry and Lua diagnostics, an opt-in density/distance test resource, and guard quadtree cleanup against lazily absent children discovered during building-pool resize.

Validated Release Win32 builds for Game SA, Multiplayer SA, and Client Deathmatch. Runtime tests exceeded the former limits with 1033 visible entities and 4677 streaming RwObject links, then cleaned up 1400 test buildings without a new crash dump.

+++ Client/game_sa/CGameSA.cpp
@@ -224,9 +224,11 @@ CGameSA::CGameSA()
         m_Pools->SetPoolCapacity(POINTER_DOUBLE_LINK_POOL, MAX_POINTER_DOUBLE_LINKS);  // Default is 3200
         dassert(m_Pools->GetPoolCapacity(POINTER_SINGLE_LINK_POOL) == MAX_POINTER_SINGLE_LINKS);
 
-        // Increase streaming object instances list size
-        MemPut<WORD>(0x05B8E55, MAX_RWOBJECT_INSTANCES * 12);  // Default is 1000 * 12
-        MemPut<WORD>(0x05B8EB0, MAX_RWOBJECT_INSTANCES * 12);  // Default is 1000 * 12
+        // GTA passes the list allocation size through two 32-bit immediates. Keep
+        // enough RwObject links for distant LOD preloading without changing when
+        // the streamer creates or removes an instance.
+        MemPut<DWORD>(0x05B8E55, MAX_RWOBJECT_INSTANCES * 12);  // Default is 1000 * 12
+        MemPut<DWORD>(0x05B8EB0, MAX_RWOBJECT_INSTANCES * 12);  // Default is 1000 * 12
 
         // Increase matrix array size
         MemPut<int>(0x054F3A1, MAX_OBJECTS * 3);  // Default is 900
